requirements.txt
atlassian-python-api==3.41.16
requests~=2.32.3
peewee~=3.17.8
python-gitlab~=5.0.0
psycopg2-binary==2.9.10
python-dateutil~=2.9.0.post0
python-dotenv==1.0.1
pyyaml==6.0.2
schedule~=1.2.2
setuptools==75.6.0
urllib3~=2.2.3